move sync structs to sync module
